fixed creation of 'available' file to handle missing .eclim dir
git-svn-id: https://eclim.svn.sourceforge.net/svnroot/eclim/trunk@2421 b5bd0a2e-1a23-0410-8567-9814cbbba304

+++ src/java/org/eclim/eclipse/AbstractEclimApplication.java
@@ -87,8 +87,10 @@ public abstract class AbstractEclimApplication
       // create marker file indicating that eclimd is up
       File marker = new File(FileUtils.concat(
             SystemUtils.USER_HOME, ".eclim/.eclimd.available"));
-      marker.createNewFile();
-      marker.deleteOnExit();
+      if(marker.getParentFile().mkdirs()){
+        marker.createNewFile();
+        marker.deleteOnExit();
+      }
 
       // start nail gun
       logger.info("Eclim Server Started on port " + port + '.');
